requirements.txt
torch
torchvision
scikit-learn
scipy 
Pillow
transformers
tensorboard
torch-tools